# APIrest/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student Management API",
    "version": "1.0.0",
    "description": "A RESTful API to manage students and their courses"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/students": {
      "get": {
        "summary": "Retrieve students (filter by id, name, email, or course using query params)",
        "tags": ["Students"],
        "parameters": [
          { "name": "id", "in": "query", "schema": { "type": "string" }, "required": false },
          { "name": "name", "in": "query", "schema": { "type": "string" }, "required": false },
          { "name": "email", "in": "query", "schema": { "type": "string", "format": "email" }, "required": false },
          { "name": "course", "in": "query", "schema": { "type": "string" }, "required": false }
        ],
        "responses": {
          "200": {
            "description": "Students retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentsResponse" },
                "example": {
                  "status": 200,
                  "message": "Students retrieved successfully",
                  "data": [
                    { "id": "1", "name": "Milton", "email": "milton@gmail.com", "courses": ["Math"] }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No students found with the given query",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "status": 404,
                  "message": "No students found with the given query",
                  "data": []
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new student",
        "tags": ["Students"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewStudent" },
              "example": {
                "name": "Jane Doe",
                "email": "jane@example.com",
                "courses": ["Math", "Science"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentResponse" },
                "example": {
                  "status": 201,
                  "message": "Student created successfully",
                  "data": {
                    "id": "2",
                    "name": "Jane Doe",
                    "email": "jane@example.com",
                    "courses": ["Math", "Science"]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "malformedJson": {
                    "summary": "Malformed JSON",
                    "value": { "status": 400, "message": "Malformed JSON in request body", "data": null }
                  },
                  "emptyBody": {
                    "summary": "Empty body",
                    "value": { "status": 400, "message": "Request body is required and must be a JSON object", "data": null }
                  },
                  "nameEmpty": {
                    "summary": "Name empty",
                    "value": { "status": 400, "message": "Name cannot be empty", "data": null }
                  },
                  "nameInvalid": {
                    "summary": "Name invalid",
                    "value": { "status": 400, "message": "Name must contain only letters and spaces", "data": null }
                  },
                  "emailEmpty": {
                    "summary": "Email empty",
                    "value": { "status": 400, "message": "Email is required", "data": null }
                  },
                  "emailInvalid": {
                    "summary": "Email invalid",
                    "value": { "status": 400, "message": "Invalid email format", "data": null }
                  },
                  "coursesInvalid": {
                    "summary": "Courses invalid",
                    "value": { "status": 400, "message": "Invalid courses format", "data": null }
                  },
                  "duplicateCourses": {
                    "summary": "Duplicate courses",
                    "value": { "status": 400, "message": "Duplicate courses are not allowed", "data": null }
                  },
                  "nonExistentCourses": {
                    "summary": "Non-existent courses",
                    "value": { "status": 400, "message": "These courses do not exist: maths 2", "data": null }
                  },
                  "invalidField": {
                    "summary": "Invalid field",
                    "value": { "status": 400, "message": "Request contains invalid field(s).", "data": null }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict (student already exists)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "status": 409,
                  "message": "Student with this email already exists",
                  "data": null
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "status": 500,
                  "message": "Internal server error",
                  "data": null
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Fully update a student by id or email (query param)",
        "tags": ["Students"],
        "parameters": [
          { "name": "id", "in": "query", "schema": { "type": "string" }, "required": false },
          { "name": "email", "in": "query", "schema": { "type": "string", "format": "email" }, "required": false }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewStudent" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing query param",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "malformedJson": {
                    "summary": "Malformed JSON",
                    "value": { "status": 400, "message": "Malformed JSON in request body", "data": null }
                  },
                  "missingQuery": {
                    "summary": "Missing query param",
                    "value": { "status": 400, "message": "'id' or 'email' is required", "data": null }
                  },
                  "emptyBody": {
                    "summary": "Empty body",
                    "value": { "status": 400, "message": "Request body cannot be empty", "data": null }
                  },
                  "nameEmpty": {
                    "summary": "Name empty",
                    "value": { "status": 400, "message": "Name cannot be empty", "data": null }
                  },
                  "nameInvalid": {
                    "summary": "Name invalid",
                    "value": { "status": 400, "message": "Name must contain only letters and spaces", "data": null }
                  },
                  "emailEmpty": {
                    "summary": "Email empty",
                    "value": { "status": 400, "message": "Email is required", "data": null }
                  },
                  "emailInvalid": {
                    "summary": "Email invalid",
                    "value": { "status": 400, "message": "Invalid email format", "data": null }
                  },
                  "coursesInvalid": {
                    "summary": "Courses invalid",
                    "value": { "status": 400, "message": "Courses must be a non-empty array of non-empty strings", "data": null }
                  },
                  "duplicateCourses": {
                    "summary": "Duplicate courses",
                    "value": { "status": 400, "message": "Duplicate courses are not allowed", "data": null }
                  },
                  "nonExistentCourses": {
                    "summary": "Non-existent courses",
                    "value": { "status": 400, "message": "These courses do not exist: maths 2", "data": null }
                  },
                  "invalidField": {
                    "summary": "Invalid field",
                    "value": { "status": 400, "message": "Request contains invalid field(s).", "data": null }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "status": 404, "message": "Student not found", "data": null }
              }
            }
          },
          "409": {
            "description": "Conflict (student already exists)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "status": 409,
                  "message": "Student with this email already exists",
                  "data": null
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "status": 500,
                  "message": "Internal server error",
                  "data": null
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Partially update a student by id or email (query param)",
        "tags": ["Students"],
        "parameters": [
          { "name": "id", "in": "query", "schema": { "type": "string" }, "required": false },
          { "name": "email", "in": "query", "schema": { "type": "string", "format": "email" }, "required": false }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "courses": {
                    "type": "array",
                    "items": { "type": "string" },
                    "minItems": 1,
                    "uniqueItems": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Student" }
                  },
                  "required": ["status", "message", "data"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing query param",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "malformedJson": {
                    "summary": "Malformed JSON",
                    "value": { "status": 400, "message": "Malformed JSON in request body", "data": null }
                  },
                  "missingQuery": {
                    "summary": "Missing query param",
                    "value": { "status": 400, "message": "'id' or 'email' is required", "data": null }
                  },
                  "emptyBody": {
                    "summary": "Empty body",
                    "value": { "status": 400, "message": "Request body cannot be empty", "data": null }
                  },
                  "nameEmpty": {
                    "summary": "Name empty",
                    "value": { "status": 400, "message": "Name cannot be empty", "data": null }
                  },
                  "nameInvalid": {
                    "summary": "Name invalid",
                    "value": { "status": 400, "message": "Name must contain only letters and spaces", "data": null }
                  },
                  "emailEmpty": {
                    "summary": "Email empty",
                    "value": { "status": 400, "message": "Email is required", "data": null }
                  },
                  "emailInvalid": {
                    "summary": "Email invalid",
                    "value": { "status": 400, "message": "Invalid email format", "data": null }
                  },
                  "coursesInvalid": {
                    "summary": "Courses invalid",
                    "value": { "status": 400, "message": "Invalid courses format", "data": null }
                  },
                  "duplicateCourses": {
                    "summary": "Duplicate courses",
                    "value": { "status": 400, "message": "Duplicate courses are not allowed", "data": null }
                  },
                  "nonExistentCourses": {
                    "summary": "Non-existent courses",
                    "value": { "status": 400, "message": "These courses do not exist: maths 2", "data": null }
                  },
                  "invalidField": {
                    "summary": "Invalid field",
                    "value": { "status": 400, "message": "Request contains invalid field(s).", "data": null }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": {}
                  },
                  "required": ["status", "message", "data"]
                },
                "example": { "status": 404, "message": "Student not found", "data": null }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a student by id or email (query param)",
        "tags": ["Students"],
        "parameters": [
          { "name": "id", "in": "query", "schema": { "type": "string" }, "required": false },
          { "name": "email", "in": "query", "schema": { "type": "string", "format": "email" }, "required": false }
        ],
        "responses": {
          "200": {
            "description": "Student deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Student" }
                  },
                  "required": ["status", "message", "data"]
                }
              }
            }
          },
          "400": {
            "description": "Missing query param",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": {}
                  },
                  "required": ["status", "message", "data"]
                },
                "example": { "status": 400, "message": "'id' or 'email' is required", "data": null }
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": {}
                  },
                  "required": ["status", "message", "data"]
                },
                "example": { "status": 404, "message": "Student not found", "data": null }
              }
            }
          }
        }
      }
    },
    "/students/by-course": {
      "get": {
        "summary": "Get students by course (query param)",
        "tags": ["Students"],
        "parameters": [
          { "name": "course", "in": "query", "schema": { "type": "string" }, "required": true }
        ],
        "responses": {
          "200": {
            "description": "Students retrieved for course",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Student" }
                    }
                  },
                  "required": ["status", "message", "data"]
                }
              }
            }
          },
          "400": {
            "description": "Missing course query param",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "integer" },
                    "message": { "type": "string" },
                    "data": {}
                  },
                  "required": ["status", "message", "data"]
                },
                "example": { "status": 400, "message": "Course required", "data": null }
              }
            }
          }
        }
      }
    },
    "/courses": {
      "post": {
        "summary": "Create a new course",
        "tags": ["Courses"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "pattern": "^[\\p{L}\\d .'-]+$",
                    "minLength": 2,
                    "example": "Mathematics"
                  }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                },
                "example": {
                  "status": 201,
                  "message": "Course created successfully",
                  "data": {
                    "course": "mathematics"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidName": {
                    "value": {
                      "status": 400,
                      "message": "Course name contains invalid characters",
                      "data": null
                    }
                  },
                  "emptyName": {
                    "value": {
                      "status": 400,
                      "message": "Course name is required",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Course already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 409,
                  "message": "Course already exists",
                  "data": null
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all courses with enrolled students",
        "tags": ["Courses"],
        "responses": {
          "200": {
            "description": "List of all courses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursesListResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Courses retrieved successfully",
                  "data": [
                    {
                      "name": "mathematics",
                      "students": ["student1@example.com", "student2@example.com"]
                    },
                    {
                      "name": "physics",
                      "students": ["student1@example.com"]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/courses/search": {
      "get": {
        "summary": "Get a specific course by name",
        "tags": ["Courses"],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Course name to search for",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Course retrieved successfully",
                  "data": {
                    "name": "mathematics",
                    "students": ["student1@example.com", "student2@example.com"]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid course name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 400,
                  "message": "Course name parameter is required",
                  "data": null
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": 404,
                  "message": "Course not found",
                  "data": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Delete a course by name",
        "description": "Deletes a course from the system based on the provided name",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Name of the course to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Course successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Course successfully deleted"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "introduction to programming"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid parameters"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Course not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Error deleting course"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Student": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "courses": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["id", "name", "email", "courses"]
      },
      "NewStudent": {
        "type": "object",
        "required": ["name", "email", "courses"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "courses": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "additionalProperties": false
      },
      "StudentsResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Student" }
          }
        },
        "required": ["status", "message", "data"]
      },
      "StudentResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/Student" }
        },
        "required": ["status", "message", "data"]
      },
      
      "CourseResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "students": {
                "type": "array",
                "items": { "type": "string", "format": "email" }
              }
            },
            "required": ["name", "students"]
          }
        },
        "required": ["status", "message", "data"]
      },
      "CoursesListResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "students": {
                  "type": "array",
                  "items": { "type": "string", "format": "email" }
                }
              },
              "required": ["name", "students"]
            }
          }
        },
        "required": ["status", "message", "data"]
      },
      "CourseStudentsResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string" },
          "data": {
            "type": "array",
            "items": { "type": "string", "format": "email" }
          }
        },
        "required": ["status", "message", "data"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer" },
          "message": { "type": "string" },
          "data": {}
        },
        "required": ["status", "message", "data"]
      }
    }
  }
}